continued porting !

+++ contracts/Migrations.sol
@@ -3,13 +3,13 @@ pragma solidity ^0.8.22;
 
 contract Migrations {
     address public owner = msg.sender;
-    uint public lastCompletedMigration;
+    uint256 public lastCompletedMigration;
 
     modifier restricted() {
         if (msg.sender == owner) _;
     }
 
-    function setCompleted(uint completed) public restricted {
+    function setCompleted(uint256 completed) public restricted {
         lastCompletedMigration = completed;
     }
